rename koiki.woocommerce.models to koiki.woocommerce.resources
