+++ koiki/create_delivery.py
@@ -1,5 +1,5 @@
 from koiki.resources import Sender, Recipient, Shipment
-from koiki.woocommerce.models import Shipping, Billing
+from koiki.woocommerce.resources import Shipping, Billing
 
 
 class CreateDelivery():
